pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "occam"
version = "0.1.0"
description = "Bayesian Occam: Experiments on concepts as modes using evidence sweeps and brittleness diagnostics"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Benji Berczi" }
]
dependencies = [
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "tqdm>=4.65.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "scipy>=1.10.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
occam = "occam.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/occam"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
